requirements-gpu.txt
tensorflow-gpu==1.13.2